Extract Title component from TitleView (UITableViewHeaderFooterView)
