+++ Sources/Mistica/Components/Title/TitleView.swift
@@ -8,77 +8,37 @@
 
 import UIKit
 
-private enum ViewStyles {
-    static let horizontalMargin: CGFloat = 16
-    static let topMargin: CGFloat = 16
-    static let bottomMargin: CGFloat = 8
-    static let minimumHeight: CGFloat = 40
-}
-
 public class TitleView: UITableViewHeaderFooterView {
-    public enum Style {
-        case title1
-        case title2
+    public typealias Style = TitleStack.Style
 
-        var font: UIFont {
-            switch self {
-            case .title1:
-                return .textPreset1(weight: .title1)
-            case .title2:
-                return .textPreset5()
-            }
-        }
-
-        var textColor: UIColor {
-            switch self {
-            case .title1:
-                return .textSecondary
-            case .title2:
-                return .textPrimary
-            }
-        }
-
-        func format(text: String?) -> String? {
-            switch self {
-            case .title1:
-                return text?.uppercased()
-            case .title2:
-                return text
-            }
-        }
-    }
-
-    private lazy var titleLabel = UILabel()
-    private lazy var linkLabel = UILabel()
+    private lazy var titleStack = TitleStack()
 
     public var onLinkLabelTapped: (() -> Void)?
 
-    public var style: Style = .default {
-        didSet {
-            updateStyle()
+    public var style: Style {
+        set {
+            titleStack.style = newValue
+        }
+        get {
+            titleStack.style
         }
     }
 
-    private var unformattedTitle: String?
     public var title: String? {
         get {
-            titleLabel.text
+            titleStack.title
         }
         set {
-            unformattedTitle = newValue
-            titleLabel.text = style.format(text: newValue)
-            layoutIfNeeded()
+            titleStack.title = newValue
         }
     }
 
     public var linkTitle: String? {
         get {
-            linkLabel.text
+            titleStack.linkTitle
         }
         set {
-            linkLabel.text = newValue
-            linkLabel.isHidden = newValue?.isEmpty ?? true
-            layoutIfNeeded()
+            titleStack.linkTitle = newValue
         }
     }
 
@@ -107,10 +67,10 @@ public class TitleView: UITableViewHeaderFooterView {
 public extension TitleView {
     var titleAccessibilityTraits: UIAccessibilityTraits {
         get {
-            titleLabel.accessibilityTraits
+            titleStack.accessibilityTraits
         }
         set {
-            titleLabel.accessibilityTraits = newValue
+            titleStack.accessibilityTraits = newValue
         }
     }
 }
@@ -119,59 +79,6 @@ public extension TitleView {
 
 private extension TitleView {
     func commonInit() {
-        layoutViews()
-        updateStyle()
-    }
-
-    func updateStyle() {
-        contentView.backgroundColor = .background
-
-        titleLabel.text = style.format(text: unformattedTitle)
-        titleLabel.font = style.font
-        titleLabel.textColor = style.textColor
-        titleLabel.numberOfLines = 0
-
-        linkLabel.font = .textPreset2(weight: .link)
-        linkLabel.textColor = .textLink
-        linkLabel.isHidden = linkLabel.text?.isEmpty ?? true
-    }
-
-    func layoutViews() {
-        linkLabel.setContentHuggingPriority(.required, for: .horizontal)
-        linkLabel.setContentCompressionResistancePriority(.required, for: .horizontal)
-        linkLabel.isUserInteractionEnabled = true
-        linkLabel.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(linkLabelTapped)))
-
-        let stackView = UIStackView(arrangedSubviews: [
-            titleLabel,
-            linkLabel
-        ])
-        stackView.spacing = 16
-        stackView.alignment = .firstBaseline
-
-        preservesSuperviewLayoutMargins = false
-        contentView.preservesSuperviewLayoutMargins = false
-        contentView.addSubview(constrainedToLayoutMarginsGuideOf: stackView)
-        contentView.directionalLayoutMargins = NSDirectionalEdgeInsets(
-            top: ViewStyles.topMargin,
-            leading: ViewStyles.horizontalMargin,
-            bottom: ViewStyles.bottomMargin,
-            trailing: ViewStyles.horizontalMargin
-        )
-    }
-
-    @objc func linkLabelTapped() {
-        onLinkLabelTapped?()
-    }
-}
-
-private extension TitleView.Style {
-    static var `default`: Self {
-        switch MisticaConfig.brandStyle {
-        case .movistar:
-            return .title2
-        case .blau, .o2, .o2New, .vivo, .custom, .vivoNew, .telefonica, .tu:
-            return .title1
-        }
+        contentView.addSubview(withDefaultConstraints: titleStack)
     }
 }
